Added Player Location Storage
The player location is now stored in the player .ini file, If the player
is registering their position is innitially set to the spawn both in the
.ini and in-game. If the player is loging in, their position is read
from the .ini file and used as spawn.
This system requires more testing and to be made a little more compact,
but it does work.

NOTE: Added Skin information to .ini but right now it is simply set to
"0" this will be fleshed out in the next commit.