Improved Login Screen
I've implimented a slightly nicer temp Login/Register screen. I've also
added a new Key to the user.ini that stores wether the user is online or
not. It serves no purpose within the .ini but it means the players data
is only stored after the player has spawned. Otherwise it is ignored.
This has fixed a huge bug which allowed the user.ini to be updated if
the player disconnected while on the login screen, this would store the
login screen health and position which is just a huge mess. This login
check solves this.

No new features have been added in-game.

Next: Simple RP commands and proximity chat.